+++ log-viewer/src/features/app/AppHeader.ts
@@ -1,13 +1,7 @@
 /*
  * Copyright (c) 2023 Certinia Inc. All rights reserved.
  */
-import {
-  provideVSCodeDesignSystem,
-  vsCodePanelTab,
-  vsCodePanelView,
-  vsCodePanels,
-} from '@vscode/webview-ui-toolkit';
-import { LitElement, css, html, unsafeCSS } from 'lit';
+import { LitElement, css, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
 import type { ApexLog } from 'apex-log-parser';
@@ -23,10 +17,7 @@ import '../find/components/FindWidget.js';
 import '../timeline/components/TimelineView.js';
 
 // styles
-import codiconStyles from '@vscode/codicons/dist/codicon.css';
 import { globalStyles } from '../../styles/global.styles.js';
-
-provideVSCodeDesignSystem().register(vsCodePanelTab(), vsCodePanelView(), vsCodePanels());
 
 @customElement('app-header')
 export class AppHeader extends LitElement {
@@ -47,7 +38,6 @@ export class AppHeader extends LitElement {
 
   static styles = [
     globalStyles,
-    unsafeCSS(codiconStyles),
     css`
       :host {
         background-color: var(--vscode-tab-activeBackground);
@@ -58,7 +48,6 @@ export class AppHeader extends LitElement {
     `,
   ];
 
-  // TODO: use @change on vscode-panels to detect tab change instead of @click on <vscode-panel-tab
   render() {
     return html`
       <nav-bar